// web/src/api/generated/CourseFacts.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CourseCode } from "./CourseCode";
import type { CourseInfo } from "./CourseInfo";

export type CourseFacts = { courses: Array<CourseInfo>, aliases: Array<[CourseCode, CourseCode]>, };